Renamed:
BackgroundTools_HTopMultilep.py -> BackgroundTools.py

as it is really not analysis-specific.
